fix floating point error

+++ Models/Calculator.cs
@@ -2,8 +2,8 @@
 
 public class Calculator
 {
-    public double Number1 { get; set; }
-    public double Number2 { get; set; }
+    public decimal Number1 { get; set; }
+    public decimal Number2 { get; set; }
     public Operation Operation { get; set; }
-    public double Result { get; set; }
+    public decimal Result { get; set; }
 }
